+++ packages/foundry-ontology/src/installation/index.ts
@@ -3,4 +3,5 @@ export {
     createFoundryOntologyRoute,
     type CreateFoundryBackendInstallationOptions,
     type FoundryConnectionOptions,
+    type FoundryOntologyRoute,
 } from "./createFoundryBackendInstallation.js";
